backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
flask-cors==4.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-slugify==8.0.4
Pillow==10.4.0
werkzeug==3.0.3
python-dotenv==1.0.1
bleach==6.1.0
